#1121 test(e2e-web): 0件スタブを utils/network へ切り出しシナリオを 1 本に集約する
